# website/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "AI Competition Platform API",
    "description": "The API documentation for the various methods that are available for querying data, setting properties or accessing info from an agent.\nQueries are cached and therefore do not update in real-time, but should always provide recent data.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0",
    "x-logo": {
      "url": "",
      "altText": "AI Competition Platform"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development"
    },
    {
      "url": "{address}:{port}",
      "description": "Custom",
      "variables": {
        "address": {
          "default": "http://localhost",
          "description": "The address of the server."
        },
        "port": {
          "default": "8080",
          "description": "The port on which the API is run."
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/games": {
      "get": {
        "summary": "Get all Games played",
        "description": "Returns all games ever played (paginated groups of 100).",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "The page of game history to request.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of games",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "games": {
                      "type": "array",
                      "maxLength": 100,
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "uuid"
                          },
                          "timePlayed": {
                            "type": "string",
                            "description": "The time the game was played.",
                            "example": "Mon Sep 08 1998 14:36:22 GMT-0700 (PDT)"
                          },
                          "isTournament": {
                            "type": "boolean",
                            "description": "If the game is a tournament game."
                          },
                          "agentScores": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "number",
                              "description": "Map of agent IDs to scores."
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Student Methods"]
      }
    },
...
    "/api/game": {
      "get": {
        "summary": "Get a Game by it's ID",
        "description": "Returns the game with the specified ID.",
        "parameters": [
          {
            "name": "gameId",
            "in": "query",
            "description": "The game ID to search for.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching games",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "games": {
                      "type": "array",
                      "maxLength": 100,
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "uuid"
                          },
                          "timePlayed": {
                            "type": "string",
                            "description": "The time the game was played.",
                            "example": "Mon Sep 08 1998 14:36:22 GMT-0700 (PDT)"
                          },
                          "gameState": {
                            "type": "string",
                            "description": "The game's log as a gzip compressed JSON. Look into 'gunzip'."
                          },
                          "agentScores": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "number",
                              "description": "Map of agent IDs to scores."
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Student Methods"]
      }
    },
    "/api/agent-games": {
      "get": {
        "summary": "Get all Games played by an Agent",
        "description": "Returns the game where the specified Agent played in.",
        "parameters": [
          {
            "name": "agentId",
            "in": "query",
            "description": "The agent ID to search for.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching games",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "games": {
                      "type": "array",
                      "maxLength": 100,
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "uuid"
                          },
                          "timePlayed": {
                            "type": "string",
                            "description": "The time the game was played.",
                            "example": "Mon Sep 08 1998 14:36:22 GMT-0700 (PDT)"
                          },
                          "gameState": {
                            "type": "string",
                            "description": "The game's log as a gzip compressed JSON. Look into 'gunzip'."
                          },
                          "agentScores": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "number",
                              "description": "Map of agent IDs to scores."
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Student Methods"]
      }
    },
    "/api/set-display-name": {
      "post": {
        "summary": "Sets your Display Name",
        "description": "Given a student number and a new name, the display name of the student number's agent is set.",
        "parameters": [
          {
            "name": "studentNumber",
            "in": "query",
            "description": "A valid user token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "displayName",
            "in": "query",
            "description": "The new display name to be set.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "error": {
                          "type": "string",
                          "example": "Error Message"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Student Methods"]
      }
    },
...
    "/api/winrate": {
      "get": {
        "summary": "Get the Win-rate of a specific Agent",
        "description": "Returns data relevant to the provided agent name.",
        "parameters": [
          {
            "name": "agentId",
            "in": "query",
            "description": "The display name of the agent to query.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The returned agent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "winrate": {
                      "type": "object",
                      "properties": {
                        "displayName": {
                          "type": "string",
                          "description": "The agent's display name.",
                          "example": "ExcitedYellowPear32"
                        },
                        "initialWinPercent": {
                          "type": "integer",
                          "description": "The original win-percentage of the agent."
                        },
                        "lastWinPercent": {
                          "type": "integer",
                          "description": "The win-percentage of the agent as of it's last game."
                        },
                        "percentageImproved": {
                          "type": "integer",
                          "description": "The percentage improvement of the agent."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Student Methods"]
      }
    },
...
    "/api/improvement": {
      "get": {
        "summary": "Get the Improvement of a specific Agent",
        "description": "Returns data relevant to the provided agent name.",
        "parameters": [
          {
            "name": "agentId",
            "in": "query",
            "description": "The display name of the agent to query.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The returned agent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "improvement": {
                      "type": "object",
                      "properties": {
                        "displayName": {
                          "type": "string",
                          "description": "The agent's display name.",
                          "example": "ExcitedYellowPear32"
                        },
                        "initialWinPercent": {
                          "type": "integer",
                          "description": "The original win-percentage of the agent."
                        },
                        "lastWinPercent": {
                          "type": "integer",
                          "description": "The win-percentage of the agent as of it's last game."
                        },
                        "percentageImproved": {
                          "type": "integer",
                          "description": "The percentage improvement of the agent."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Student Methods"]
      }
    },
...
    "/api/authentication": {
      "post": {
        "summary": "Get the Available Games",
        "description": "Returns a list of available games.",
        "parameters": [
          {
            "name": "agentToken",
            "in": "query",
            "description": "The student token to verify.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If the token is valid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authenticated": {
                      "type": "boolean",
                      "description": "If the token is valid."
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Student Methods"]
      }
    },
    "/api/check-admin": {
      "get": {
        "summary": "Validate an Admin Token",
        "description": "Returns a success if the provided token is an admin token.",
        "parameters": [
          {
            "name": "adminToken",
            "in": "query",
            "description": "The admin token to verify.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If the token is valid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authenticated": {
                      "type": "boolean",
                      "description": "If the token is valid or not."
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Admin Methods"]
      }
    },
    "/api/admin-view": {
      "get": {
        "summary": "Get a list of all Users",
        "description": "Returns a list of all users, showing their student number, display name and token.",
        "parameters": [
          {
            "name": "adminToken",
            "in": "query",
            "description": "Admin token required.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "object",
                      "properties": {
                        "studentNumber": {
                          "type": "string",
                          "description": "The student number of this user."
                        },
                        "displayName": {
                          "type": "string",
                          "description": "The display name of this user.",
                          "example": "ExcitedYellowPear32"
                        },
                        "authToken": {
                          "type": "string",
                          "description": "The authentication token of this user."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Admin Methods"]
      }
    },
    "/api/generate-token": {
      "post": {
        "summary": "Create a new Student Token",
        "description": "Creates a new student token, given a seed input and authorization.",
        "parameters": [
          {
            "name": "adminToken",
            "in": "query",
            "description": "Admin token required.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "description": "The seed to generate the token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The generated token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "The generated token."
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Admin Methods"]
      }
    },
    "/api/generate-admin-token": {
      "post": {
        "summary": "Create a new Admin Token",
        "description": "Creates a new student token, given a seed input and authorization.",
        "parameters": [
          {
            "name": "adminToken",
            "in": "query",
            "description": "Admin token required.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "description": "The seed to generate the token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The generated token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "The generated token."
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Admin Methods"]
      }
    },
    "/api/admin-get-tournament-winrate": {
      "get": {
        "summary": "Get the Agents with the Top Win-Rates in Tournament Games",
        "description": "Returns an array of data corresponding to the top agents. (Warning, this query may lock up the server, hence admin auth required.)",
        "parameters": [
          {
            "name": "adminToken",
            "in": "query",
            "description": "Admin token required.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "winrate": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "displayName": {
                            "type": "string",
                            "description": "The agent's display name.",
                            "example": "ExcitedYellowPear32"
                          },
                          "isBot": {
                            "type": "boolean",
                            "description": "If the agent is a bot."
                          },
                          "gamesPlayed": {
                            "type": "integer",
                            "description": "How many games the agent has played."
                          },
                          "wins": {
                            "type": "integer",
                            "description": "How many games the agent has won."
                          },
                          "winPercent": {
                            "type": "number",
                            "description": "The agent's percentage win-rate."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Admin Methods"]
      }
    },
    "/api/started": {
      "get": {
        "summary": "Returns if the Game has Started",
        "description": "Returns a boolean indicating if the game has begun for an agent.",
        "parameters": [
          {
            "name": "agentToken",
            "in": "query",
            "description": "The token belonging to the agent who is checking.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If the game has started.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "gameStarted": {
                      "type": "boolean",
                      "description": "If the game has started."
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Agent Methods"]
      }
    },
...
  }
}